Split peel_depth_fragment shader into 4 (D0, D1, D2, D3) to avoid read/write error
Changes were made in the WebGL specs to check for reading and writing
to the same texture buffer in one drawElements pass. Because there were
if statements in the shader, to determine whether a texture should be
written or read in this pass, the new shader language compiler saw
both reads and writes of the same texture buffer, which looked like
an error. A side effect of this change is that transparency comes
closer to work on mobile devices: the z ordering is now correct and
includes all peels, but the transparent objects have ugly bands.