Reorient ring texture seams and direction
As with other shapes, hide the texture seams toward the back
for ring (and compound ring).  Also reorient the texture so that
the "equator" is along the outer ring (in yz-plane in model coords).

Also, ring triangles previously had the wrong handedness!
(Doesn't seem to matter unless culling is on, in which case the rings
looked super trippy inside-out.)