Comment makeRing_compound more
Describe in more detail how the elliptical ring surface points are
generated, as well as the normals and bumpaxes.

Also document explicitly that the calculated normals are approximate.
(They're exact for a circular ring, and get worse with eccentricity.
They're least accurate at phi=pi/4 and related points, and less accurate
on the inner edge than the outer edge of the ring.  For lighting
purposes they'll work OK.  Calculating the exact result would be
tedious.)